mkzfs: warn about non-empty /zfs
Alternatively we could just remove it, but doing recursive removes in
a top-level directory as root seem a bit too dangerous for a build
process.

+++ scripts/mkzfs.py
@@ -50,6 +50,10 @@ dev='/dev/vblk1'
 zfs_pool='osv'
 zfs_fs='usr'
 
+if os.path.exists(zfs_root) and os.listdir(zfs_root): 
+    print 'Please make sure %s does not exist or is an empty directory' % zfs_root
+    sys.exit(1)
+
 os.system('sudo mkdir -p %s' % zfs_root)
 
 os.system('sudo rm -f %s' % options.output)
